Get switch result for tkonfound based on chi-square tests

This function serves for tkonfound

getswitch_chisq(a, b, c, d, thr_p = 0.05, switch_trm = T)

Arguments

a
cell is the number of cases in the control group showing unsuccessful results

b
cell is the number of cases in the control group showing successful results

c
cell is the number of cases in the treatment group showing unsuccessful results

d
cell is the number of cases in the treatment group showing successful results

thr_p
the p-value threshold used to evaluate statistical significance, with the default of 0.05

switch_trm
whether switching the two cells in the treatment row or the two cells in the control row, with the default of the treatment row

Value

prints the results that are needed for tkonfound

Examples

# using getswitch_chisq for a 2x2 table
getswitch_chisq(35, 17, 17, 38)

#> $final_switch
#> [1] 10
#> 
#> $User_enter_value
#>      [,1] [,2]
#> [1,]   35   17
#> [2,]   17   38
#> 
#> $Transfer_Table
#>      [,1] [,2]
#> [1,]   35   17
#> [2,]   27   28
#> 
#> $p_final
#> [1] 0.05641045
#> 
#> $chisq_final
#> X-squared 
#>  3.639896 
#> 
#> $needtworows
#> [1] FALSE
#> 
#> $taylor_pred
#> [1] 9.520577
#> 
#> $perc_bias_pred
#> [1] 8.504605
#> 
#> $final_extra
#> [1] 0
#> 
#> $dcroddsratio_ob
#> [1] TRUE
#> 
#> $total_switch
#> [1] 10
#> 
#> $isinvalidate_ob
#> [1] TRUE
#> 

getswitch_chisq(35, 17, 17, 38, thr_p = 0.01)

#> $final_switch
#> [1] 7
#> 
#> $User_enter_value
#>      [,1] [,2]
#> [1,]   35   17
#> [2,]   17   38
#> 
#> $Transfer_Table
#>      [,1] [,2]
#> [1,]   35   17
#> [2,]   24   31
#> 
#> $p_final
#> [1] 0.01386832
#> 
#> $chisq_final
#> X-squared 
#>  6.054828 
#> 
#> $needtworows
#> [1] FALSE
#> 
#> $taylor_pred
#> [1] 5.948128
#> 
#> $perc_bias_pred
#> [1] 5.286495
#> 
#> $final_extra
#> [1] 0
#> 
#> $dcroddsratio_ob
#> [1] TRUE
#> 
#> $total_switch
#> [1] 7
#> 
#> $isinvalidate_ob
#> [1] TRUE
#> 

getswitch_chisq(35, 17, 17, 38, thr_p = 0.01, switch_trm = FALSE)

#> $final_switch
#> [1] 7
#> 
#> $User_enter_value
#>      [,1] [,2]
#> [1,]   35   17
#> [2,]   17   38
#> 
#> $Transfer_Table
#>      [,1] [,2]
#> [1,]   28   24
#> [2,]   17   38
#> 
#> $p_final
#> [1] 0.01629644
#> 
#> $chisq_final
#> X-squared 
#>  5.770603 
#> 
#> $needtworows
#> [1] FALSE
#> 
#> $taylor_pred
#> [1] 5.694875
#> 
#> $perc_bias_pred
#> [1] 5.150052
#> 
#> $final_extra
#> [1] 0
#> 
#> $dcroddsratio_ob
#> [1] TRUE
#> 
#> $total_switch
#> [1] 7
#> 
#> $isinvalidate_ob
#> [1] TRUE
#>