Perform meta-analyses including sensitivity analysis

For fitted models, this command carries out sensitivity analysis for a number of models, when their parameters stored in a data.frame.

mkonfound(d, t, df, alpha = 0.05, tails = 2, return_plot = FALSE)

Arguments

d
data.frame or tibble with the t-statistics and associated degrees of freedom

t
t-statistic or vector of t-statistics

df
degrees of freedom or vector of degrees of freedom associated with the t-statistics in the t argument

alpha
probability of rejecting the null hypothesis (defaults to 0.05)

tails
integer whether hypothesis testing is one-tailed (1) or two-tailed (2; defaults to 2)

return_plot
whether to return a plot of the percent bias; defaults to FALSE

Value

prints the bias and the number of cases that would have to be replaced with cases for which there is no effect to invalidate the inference for each of the cases in the data.frame

Examples

d <- read.csv("https://msu.edu/~kenfrank/example%20dataset%20for%20mkonfound.csv")
str(d)

#> 'data.frame':	30 obs. of  2 variables:
#>  $ t : num  7.08 4.13 1.89 -4.17 -1.19 ...
#>  $ df: int  178 193 47 138 97 87 117 75 137 195 ...

mkonfound(d, t, df)

#> # A tibble: 30 x 7
#>            t    df        action           inference
#>        <dbl> <int>         <chr>               <chr>
#>  1  7.076763   178 to_invalidate         reject_null
#>  2  4.127893   193 to_invalidate         reject_null
#>  3  1.893137    47    to_sustain fail_to_reject_null
#>  4 -4.166395   138 to_invalidate         reject_null
#>  5 -1.187599    97    to_sustain fail_to_reject_null
#>  6  3.585478    87 to_invalidate         reject_null
#>  7  0.281938   117    to_sustain fail_to_reject_null
#>  8  2.549647    75 to_invalidate         reject_null
#>  9 -4.436048   137 to_invalidate         reject_null
#> 10 -2.045373   195 to_invalidate         reject_null
#> # ... with 20 more rows, and 3 more variables:
#> #   pct_bias_to_change_inference <dbl>, itcv <dbl>, r_con <dbl>