Perform meta-analyses including sensitivity analysis

For fitted models, this command carries out sensitivity analysis for a number of models, when their parameters stored in a data.frame.

mkonfound_d(df, alpha = 0.05, tails = 2, return_plot = FALSE,
  component_correlations = FALSE)

Arguments

df
a data.frame with columns for the 'unstd_beta', 'std_err', 'n_obs', and 'n_covariates' arguments (see ?pkonfound for more information on these)

alpha
probability of rejecting the null hypothesis (defaults to 0.05)

tails
integer whether hypothesis testing is one-tailed (1) or two-tailed (2; defaults to 2)

return_plot
whether to return a plot of the percent bias; defaults to FALSE

component_correlations
whether to return the component correlations as part of the correlation-based approach

Value

prints the bias and the number of cases that would have to be replaced with cases for which there is no effect to invalidate the inference for each of the cases in the data.frame

Examples

df <- data.frame(unstd_beta = c(2, 10, 1.7, .4, 3.2, 1.0, 2.3, 4.1, .9),
                 std_error = c(.3, 2.9, 1.5, 2, .1, .04, .1, 3, .5),
                 n_obs = c(70, 405, 200, 100, 103, 20, 50, 721, 320),
                 n_covs = c(3, 4, 1, 3, 10, 4, 1, 1, 0))
mkonfound(df)

#> Error: `var` must be a value between 0 and 0 (excluding zero), not -1

mkonfound(df, return_plot = TRUE)

#> Error: `var` must be a value between 0 and 0 (excluding zero), not -1