Plot Effect Threshold Diagram

This function creates a plot to illustrate the threshold of an effect estimate in relation to a specified beta threshold. It uses ggplot2 for graphical representation.

Usage

plot_threshold(beta_threshold, est_eff)

Arguments

beta_threshold

The threshold value for the effect.

est_eff

The estimated effect size.

Value

A ggplot object representing the effect threshold diagram.