Perform sensitivity analysis for a 2x2 table

This function calculates the number of cases (RIS) that would have to be switched from one cell to another of a 2x2 table to invalidate an inference made about the association between the rows and columns. This can be applied to treatment vs control with successful vs unsuccessful outcomes.

tkonfound(a, b, c, d, thr_p = 0.05, switch_trm = T)

Arguments

a
cell is the number of cases in the control group showing unsuccessful results

b
cell is the number of cases in the control group showing successful results

c
cell is the number of cases in the treatment group showing unsuccessful results

d
cell is the number of cases in the treatment group showing successful results

thr_p
the p-value threshold used to evaluate statistical significance, with the default of 0.05

switch_trm
whether switching the two cells in the treatment row or the two cells in the control row, with the default of the treatment row

Value

prints a 2x2 table after switching cases from one cell to another so that the inference is invalidated about the association between the rows and columns

Examples

# using tkonfound for a 2x2 table
tkonfound(35, 17, 17, 38)

#> Warning: 'x' has been rounded to integer: Mean relative difference: 1

#> [[1]]
#> [1] "The tkonfound function calculates the number of cases that would have to be replaced with no effect cases (RIR) to invalidate an inference made about the association between the rows and columns in a 2x2 table. One can also interpret this as switches from one cell to another, such as from the treatment success cell to the treatment failure cell."
#> 
#> [[2]]
#> [1] "See konfound_fig for full and accessible details in graphic form!"
#> 
#> [[3]]
#> [1] "To invalidate the inference, you would need to replace 13 treatment success cases with null hypothesis cases (RIR = 13). This is equivalent to transferring 9 cases from treatment success to treatment failure ."
#> 
#> $User_enter_value
#>           Fail Success
#> Control     35      17
#> Treatment   17      38
#> 
#> $Transfer_Table
#>           Fail Success
#> Control     35      17
#> Treatment   26      29
#> 
#> [[6]]
#> [1] "For the User-entered Table, we have an estimated odds ratio of 4.530, with p value of 0.000."
#> 
#> [[7]]
#> [1] "For the Transfer Table, we have an estimated odds ratio of 2.278, with p value of 0.051."
#> 
#> $RIR
#> [1] 13
#> 

tkonfound(35, 17, 17, 38, thr_p = 0.01)

#> Warning: 'x' has been rounded to integer: Mean relative difference: 1

#> [[1]]
#> [1] "The tkonfound function calculates the number of cases that would have to be replaced with no effect cases (RIR) to invalidate an inference made about the association between the rows and columns in a 2x2 table. One can also interpret this as switches from one cell to another, such as from the treatment success cell to the treatment failure cell."
#> 
#> [[2]]
#> [1] "See konfound_fig for full and accessible details in graphic form!"
#> 
#> [[3]]
#> [1] "To invalidate the inference, you would need to replace 9 treatment success cases with null hypothesis cases (RIR = 9). This is equivalent to transferring 6 cases from treatment success to treatment failure ."
#> 
#> $User_enter_value
#>           Fail Success
#> Control     35      17
#> Treatment   17      38
#> 
#> $Transfer_Table
#>           Fail Success
#> Control     35      17
#> Treatment   23      32
#> 
#> [[6]]
#> [1] "For the User-entered Table, we have an estimated odds ratio of 4.530, with p value of 0.000."
#> 
#> [[7]]
#> [1] "For the Transfer Table, we have an estimated odds ratio of 2.835, with p value of 0.011."
#> 
#> $RIR
#> [1] 9
#> 

tkonfound(35, 17, 17, 38, thr_p = 0.01, switch_trm = FALSE)

#> Warning: 'x' has been rounded to integer: Mean relative difference: 1

#> [[1]]
#> [1] "The tkonfound function calculates the number of cases that would have to be replaced with no effect cases (RIR) to invalidate an inference made about the association between the rows and columns in a 2x2 table. One can also interpret this as switches from one cell to another, such as from the treatment success cell to the treatment failure cell."
#> 
#> [[2]]
#> [1] "See konfound_fig for full and accessible details in graphic form!"
#> 
#> [[3]]
#> [1] "To invalidate the inference, you would need to replace 18 control failure cases with null hypothesis cases (RIR = 18). This is equivalent to transferring 6 cases from control failure to control success ."
#> 
#> $User_enter_value
#>           Fail Success
#> Control     35      17
#> Treatment   17      38
#> 
#> $Transfer_Table
#>           Fail Success
#> Control     29      23
#> Treatment   17      38
#> 
#> [[6]]
#> [1] "For the User-entered Table, we have an estimated odds ratio of 4.530, with p value of 0.000."
#> 
#> [[7]]
#> [1] "For the Transfer Table, we have an estimated odds ratio of 2.790, with p value of 0.012."
#> 
#> $RIR
#> [1] 18
#>