Perform sensitivity analysis on fitted models

For fitted models, this command calculates (1) how much bias there must be in an estimate to invalidate/sustain an inference; (2) the impact of an omitted variable necessary to invalidate/sustain an inference for a regression coefficient. Currently works for: models created with lm() (linear models).

konfound(model_object, tested_variable, alpha = 0.05, tails = 2,
  to_return = "print", test_all = FALSE, component_correlations = F)

Arguments

model_object
output from a model (currently works for: lm)

tested_variable
Variable associated with the unstandardized beta coefficient to be tested

alpha
probability of rejecting the null hypothesis (defaults to 0.05)

tails
integer whether hypothesis testing is one-tailed (1) or two-tailed (2; defaults to 2)

to_return
whether to return a data.frame (by specifying this argument to equal "df"), table ("table"), or a plot ("plot"); default is to print the output to the console

test_all
whether to carry out the sensitivity test for all of the coefficients (defaults to FALSE)

component_correlations
whether to return the component correlations as part of the correlation-based approach

Value

prints the bias and the number of cases that would have to be replaced with cases for which there is no effect to invalidate the inference

Examples

m1 <- lm(mpg ~ wt + hp, data = mtcars)
konfound(m1, wt)

#> Replacement of Cases Approach:
#> To invalidate the inference, 66.664 % of the estimate would have to be due to bias.To invalidate the inference, 21 observations would have to be replaced with cases for which the effect is 0.
#> 
#> Correlation-based Approach:
#> An omitted variable would have to be correlated at 0.787 with the outcome and at 0.787 with the predictor of interest (conditioning on observed covariates) to invalidate an inference.
#> Correspondingly the impact of an omitted variable (as defined in Frank 2000) must be 0.787 X 0.787 = 0.619 to invalidate an inference.

konfound(m1, wt, test_all = TRUE)

#>   var_name unstd_beta std_err n_obs n_covariates     inference percent_bias
#> 1       wt     -3.878   0.633    32            1 to_invalidate       66.664
#> 2       hp     -0.032   0.009    32            1 to_invalidate       42.561
#>   replace_null_cases beta_threshhold omitted_variable_corr       itcv
#> 1                 21     -1.29275846                 0.787 -0.6199455
#> 2                 14     -0.01838045                 0.556 -0.3088738

konfound(m1, wt, to_return = "table")

#> Dependent variable is mpg 

#>          term estimate std.error statistic p.value  itcv impact
#> 1 (Intercept)   37.227     1.599    23.285   0.000    NA     NA
#> 2          wt   -3.878     0.633    -6.129   0.000 0.243     NA
#> 3          hp   -0.032     0.009    -3.519   0.001    NA  0.511