Perform sensitivity analysis for a 2x2 table

This function calculates the number of cases (RIS) that would have to be switched from one cell to another of a 2x2 table to invalidate an inference made about the association between the rows and columns. This can be applied to treatment vs control with successful vs unsuccessful outcomes.

tkonfound(a, b, c, d, thr_p = 0.05, switch_trm = T)

Arguments

a
cell is the number of cases in the control group showing unsuccessful results

b
cell is the number of cases in the control group showing successful results

c
cell is the number of cases in the treatment group showing unsuccessful results

d
cell is the number of cases in the treatment group showing successful results

thr_p
the p-value threshold used to evaluate statistical significance, with the default of 0.05

switch_trm
whether switching the two cells in the treatment row or the two cells in the control row, with the default of the treatment row

Value

prints a 2x2 table after switching cases from one cell to another so that the inference is invalidated about the association between the rows and columns

Examples

# using tkonfound for a 2x2 table
tkonfound(35, 17, 17, 38)

#> [[1]]
#> [1] "The tkonfound function calculates the number of cases that would have to be switched from one cell to another of a 2x2 table to invalidate an inference made about the association between the rows and columns. This can be applied to treatment vs control with successful vs unsuccessful outcomes."
#> 
#> [[2]]
#> [1] "See konfound_fig for full and accessible details in graphic form!"
#> 
#> [[3]]
#> [1] "To invalidate the inference, 10 cases need to be transferred from treatment success to treatment failure, as shown, from the User-entered Table to the Transfer Table."
#> 
#> $User_enter_value
#>           Fail Success
#> Control     35      17
#> Treatment   17      38
#> 
#> $Transfer_Table
#>           Fail Success
#> Control     35      17
#> Treatment   27      28
#> 
#> [[6]]
#> [1] "For the User-entered Table, we have a Pearson's chi square of 14.176, with p value of 0.000."
#> 
#> [[7]]
#> [1] "For the Transfer Table, we have a Pearson's chi square of 3.640, with p value of 0.056."
#> 

tkonfound(35, 17, 17, 38, thr_p = 0.01)

#> [[1]]
#> [1] "The tkonfound function calculates the number of cases that would have to be switched from one cell to another of a 2x2 table to invalidate an inference made about the association between the rows and columns. This can be applied to treatment vs control with successful vs unsuccessful outcomes."
#> 
#> [[2]]
#> [1] "See konfound_fig for full and accessible details in graphic form!"
#> 
#> [[3]]
#> [1] "To invalidate the inference, 7 cases need to be transferred from treatment success to treatment failure, as shown, from the User-entered Table to the Transfer Table."
#> 
#> $User_enter_value
#>           Fail Success
#> Control     35      17
#> Treatment   17      38
#> 
#> $Transfer_Table
#>           Fail Success
#> Control     35      17
#> Treatment   24      31
#> 
#> [[6]]
#> [1] "For the User-entered Table, we have a Pearson's chi square of 14.176, with p value of 0.000."
#> 
#> [[7]]
#> [1] "For the Transfer Table, we have a Pearson's chi square of 6.055, with p value of 0.014."
#> 

tkonfound(35, 17, 17, 38, thr_p = 0.01, switch_trm = FALSE)

#> [[1]]
#> [1] "The tkonfound function calculates the number of cases that would have to be switched from one cell to another of a 2x2 table to invalidate an inference made about the association between the rows and columns. This can be applied to treatment vs control with successful vs unsuccessful outcomes."
#> 
#> [[2]]
#> [1] "See konfound_fig for full and accessible details in graphic form!"
#> 
#> [[3]]
#> [1] "To invalidate the inference, 7 cases need to be transferred from control failure to control success, as shown, from the User-entered Table to the Transfer Table."
#> 
#> $User_enter_value
#>           Fail Success
#> Control     35      17
#> Treatment   17      38
#> 
#> $Transfer_Table
#>           Fail Success
#> Control     28      24
#> Treatment   17      38
#> 
#> [[6]]
#> [1] "For the User-entered Table, we have a Pearson's chi square of 14.176, with p value of 0.000."
#> 
#> [[7]]
#> [1] "For the Transfer Table, we have a Pearson's chi square of 5.771, with p value of 0.016."
#>